figured out entry point pkh needs to be main to be able to be run
