+++ .gitignore
@@ -1,11 +1,8 @@
 /vendor
 /.idea
-Homestead.json
-Homestead.yaml
-.env
-.phpunit.result.cache
 node_modules
 public/*
 main
 assets
 database
+mokintoken
